Fixed lport bug

+++ kernel/defs.h
@@ -70,7 +70,7 @@ void            net_tx_udp(struct mbuf*, uint32, uint16, uint16);
 
 // sysnet.c
 void            sockinit(void);
-int             sockalloc(struct file **, uint32, uint16, uint16, uint8);
+int             sockalloc(struct file **, uint32, uint32, uint32, uint8);
 void            sockclose(struct sock*, int);
 int             sockread(struct sock*, uint64, int);
 int             sockwrite(struct sock*, uint64, int);
